Refactor constants into constants submodule, update tests
