+++ fastapi_mvc/__init__.py
@@ -2,8 +2,6 @@
 import logging
 
 from fastapi_mvc.core import (
-    ANSWERS_FILE,
-    VERSION,
     ClickAliasedGroup,
     Command,
     Generator,
@@ -16,8 +14,6 @@ log.addHandler(logging.NullHandler())
 
 
 __all__ = (
-    "ANSWERS_FILE",
-    "VERSION",
     "ClickAliasedGroup",
     "Command",
     "Generator",
